Needs checking: LibreQoS.py will select .insight files if they are available and insight management is enabled.

+++ src/rust/lqos_python/src/lib.rs
@@ -400,13 +400,13 @@ fn run_shell_commands_as_sudo() -> PyResult<bool> {
 #[pyfunction]
 fn generated_pn_download_mbps() -> PyResult<u32> {
     let config = lqos_config::load_config().unwrap();
-    Ok(config.queues.generated_pn_download_mbps)
+    Ok(config.queues.generated_pn_download_mbps as u32)
 }
 
 #[pyfunction]
 fn generated_pn_upload_mbps() -> PyResult<u32> {
     let config = lqos_config::load_config().unwrap();
-    Ok(config.queues.generated_pn_upload_mbps)
+    Ok(config.queues.generated_pn_upload_mbps as u32)
 }
 
 #[pyfunction]
